schnorr: Move sig code to signature files.
This moves the signature code from ecdsa.go and ecdsa_test.go to
signature.go and signature_test.go where it more properly belongs since
not only is it dealing with signatures, ECDSA is another algorithm
altogether and thus it doesn't make sense to name the files that.
