--- dcrec/secp256k1/schnorr/ecdsa.go
@@ -1,298 +1,0 @@
-// Copyright (c) 2015-2020 The Decred developers
-// Use of this source code is governed by an ISC
-// license that can be found in the LICENSE file.
-
-package schnorr
-
-import (
-	"bytes"
-	"crypto/ecdsa"
-	"fmt"
-	"io"
-	"math/big"
-
-	"github.com/decred/dcrd/chaincfg/chainhash"
-	"github.com/decred/dcrd/dcrec/secp256k1/v3"
-)
-
-// scalarSize is the size of an encoded big endian scalar.
-const scalarSize = 32
-
-var (
-	// bigZero is the big representation of zero.
-	bigZero = new(big.Int).SetInt64(0)
-)
-
-// zeroArray zeroes the memory of a scalar array.
-func zeroArray(a *[scalarSize]byte) {
-	for i := 0; i < scalarSize; i++ {
-		a[i] = 0x00
-	}
-}
-
-// zeroSlice zeroes the memory of a scalar byte slice.
-func zeroSlice(s []byte) {
-	for i := 0; i < scalarSize; i++ {
-		s[i] = 0x00
-	}
-}
-
-// zeroBigInt zeroes the underlying memory used by the passed big integer.  The
-// big integer must not be used after calling this as it changes the internal
-// state out from under it which can lead to unpredictable results.
-func zeroBigInt(v *big.Int) {
-	words := v.Bits()
-	for i := 0; i < len(words); i++ {
-		words[i] = 0
-	}
-	v.SetInt64(0)
-}
-
-// GenerateKey generates a key using a random number generator, returning
-// the private scalar and the corresponding public key points.
-func GenerateKey(rand io.Reader) (priv []byte, x, y *big.Int, err error) {
-	key, err := ecdsa.GenerateKey(secp256k1.S256(), rand)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-	return key.D.Bytes(), key.PublicKey.X, key.PublicKey.Y, nil
-}
-
-// schnorrSign signs a Schnorr signature using a specified hash function
-// and the given nonce, private key, message, and optional public nonce.
-// CAVEAT: Lots of variable time algorithms using both the private key and
-// k, which can expose the signer to constant time attacks. You have been
-// warned! DO NOT use this algorithm where you might have the possibility
-// of someone having EM field/cache/etc access.
-// Memory management is also kind of sloppy and whether or not your keys
-// or nonces can be found in memory later is likely a product of when the
-// garbage collector runs.
-// TODO Use field elements with constant time algorithms to prevent said
-// attacks.
-// This is identical to the Schnorr signature function found in libsecp256k1:
-// https://github.com/bitcoin/secp256k1/tree/master/src/modules/schnorr
-func schnorrSign(msg []byte, ps []byte, k []byte,
-	pubNonceX *big.Int, pubNonceY *big.Int,
-	hashFunc func([]byte) []byte) (*Signature, error) {
-	curve := secp256k1.S256()
-	if len(msg) != scalarSize {
-		str := fmt.Sprintf("wrong size for message (got %v, want %v)",
-			len(msg), scalarSize)
-		return nil, schnorrError(ErrBadInputSize, str)
-	}
-	if len(ps) != scalarSize {
-		str := fmt.Sprintf("wrong size for privkey (got %v, want %v)",
-			len(ps), scalarSize)
-		return nil, schnorrError(ErrBadInputSize, str)
-	}
-	if len(k) != scalarSize {
-		str := fmt.Sprintf("wrong size for nonce k (got %v, want %v)",
-			len(k), scalarSize)
-		return nil, schnorrError(ErrBadInputSize, str)
-	}
-
-	psBig := new(big.Int).SetBytes(ps)
-	bigK := new(big.Int).SetBytes(k)
-
-	if psBig.Cmp(bigZero) == 0 {
-		str := fmt.Sprintf("secret scalar is zero")
-		return nil, schnorrError(ErrInputValue, str)
-	}
-	if psBig.Cmp(curve.N) >= 0 {
-		str := fmt.Sprintf("secret scalar is out of bounds")
-		return nil, schnorrError(ErrInputValue, str)
-	}
-	if bigK.Cmp(bigZero) == 0 {
-		str := fmt.Sprintf("k scalar is zero")
-		return nil, schnorrError(ErrInputValue, str)
-	}
-	if bigK.Cmp(curve.N) >= 0 {
-		str := fmt.Sprintf("k scalar is out of bounds")
-		return nil, schnorrError(ErrInputValue, str)
-	}
-
-	// R = kG
-	var Rpx, Rpy *big.Int
-	Rpx, Rpy = curve.ScalarBaseMult(k)
-	if pubNonceX != nil && pubNonceY != nil {
-		// Optional: if k' exists then R = R+k'
-		Rpx, Rpy = curve.Add(Rpx, Rpy, pubNonceX, pubNonceY)
-	}
-
-	// Check if the field element that would be represented by Y is odd.
-	// If it is, just keep k in the group order.
-	if Rpy.Bit(0) == 1 {
-		bigK.Mod(bigK, curve.N)
-		bigK.Sub(curve.N, bigK)
-	}
-
-	// h = Hash(r || m)
-	Rpxb := bigIntToEncodedBytes(Rpx)
-	hashInput := make([]byte, 0, scalarSize*2)
-	hashInput = append(hashInput, Rpxb[:]...)
-	hashInput = append(hashInput, msg...)
-	h := hashFunc(hashInput)
-	hBig := new(big.Int).SetBytes(h)
-
-	// If the hash ends up larger than the order of the curve, abort.
-	if hBig.Cmp(curve.N) >= 0 {
-		str := fmt.Sprintf("hash of (R || m) too big")
-		return nil, schnorrError(ErrSchnorrHashValue, str)
-	}
-
-	// s = k - hx
-	// TODO Speed this up a bunch by using field elements, not
-	// big ints. That we multiply the private scalar using big
-	// ints is also probably bad because we can only assume the
-	// math isn't in constant time, thus opening us up to side
-	// channel attacks. Using a constant time field element
-	// implementation will fix this.
-	sBig := new(big.Int)
-	sBig.Mul(hBig, psBig)
-	sBig.Sub(bigK, sBig)
-	sBig.Mod(sBig, curve.N)
-
-	if sBig.Cmp(bigZero) == 0 {
-		str := fmt.Sprintf("sig s %v is zero", sBig)
-		return nil, schnorrError(ErrZeroSigS, str)
-	}
-
-	// Zero out the private key and nonce when we're done with it.
-	zeroBigInt(bigK)
-	zeroSlice(k)
-	zeroBigInt(psBig)
-	zeroSlice(ps)
-
-	return &Signature{Rpx, sBig}, nil
-}
-
-// nonceRFC6979 is a local instantiation of deterministic nonce generation
-// by the standards of RFC6979.
-func nonceRFC6979(privKey []byte, hash []byte, extra []byte, version []byte, extraIterations uint32) []byte {
-	k := secp256k1.NonceRFC6979(privKey, hash, extra, version, extraIterations)
-	kBytes := k.Bytes()
-	defer zeroArray(&kBytes)
-	bigK := new(big.Int).SetBytes(kBytes[:])
-	defer zeroBigInt(bigK)
-	nonce := bigIntToEncodedBytes(bigK)
-	return nonce[:]
-}
-
-// Sign is the exported version of sign. It uses RFC6979 and Blake256 to
-// produce a Schnorr signature.
-func Sign(priv *secp256k1.PrivateKey, hash []byte) (r, s *big.Int, err error) {
-	// Convert the private scalar to a 32 byte big endian number.
-	bigPriv := new(big.Int).SetBytes(priv.Serialize())
-	pA := bigIntToEncodedBytes(bigPriv)
-	defer zeroArray(pA)
-
-	for iteration := uint32(0); ; iteration++ {
-		// Generate a 32-byte scalar to use as a nonce via RFC6979.
-		kB := nonceRFC6979(priv.Serialize(), hash, nil, nil, iteration)
-		sig, err := schnorrSign(hash, pA[:], kB, nil, nil, chainhash.HashB)
-		if err == nil {
-			r = sig.r
-			s = sig.s
-			break
-		}
-
-		errTyped, ok := err.(Error)
-		if !ok {
-			return nil, nil, fmt.Errorf("unknown error type")
-		}
-		if errTyped.ErrorCode != ErrSchnorrHashValue {
-			return nil, nil, err
-		}
-	}
-
-	return r, s, nil
-}
-
-// schnorrVerify is the internal function for verification of a secp256k1
-// Schnorr signature. A secure hash function may be passed for the calculation
-// of r.
-// This is identical to the Schnorr verification function found in libsecp256k1:
-// https://github.com/bitcoin/secp256k1/tree/master/src/modules/schnorr
-func schnorrVerify(sig []byte,
-	pubkey *secp256k1.PublicKey, msg []byte, hashFunc func([]byte) []byte) (bool,
-	error) {
-	curve := secp256k1.S256()
-	if len(msg) != scalarSize {
-		str := fmt.Sprintf("wrong size for message (got %v, want %v)",
-			len(msg), scalarSize)
-		return false, schnorrError(ErrBadInputSize, str)
-	}
-
-	if len(sig) != SignatureSize {
-		str := fmt.Sprintf("wrong size for signature (got %v, want %v)",
-			len(sig), SignatureSize)
-		return false, schnorrError(ErrBadInputSize, str)
-	}
-	if pubkey == nil {
-		str := fmt.Sprintf("nil pubkey")
-		return false, schnorrError(ErrInputValue, str)
-	}
-
-	if !curve.IsOnCurve(pubkey.X(), pubkey.Y()) {
-		str := fmt.Sprintf("pubkey point is not on curve")
-		return false, schnorrError(ErrPointNotOnCurve, str)
-	}
-
-	sigR := sig[:32]
-	sigS := sig[32:]
-	sigRCopy := make([]byte, scalarSize)
-	copy(sigRCopy, sigR)
-	toHash := append(sigRCopy, msg...)
-	h := hashFunc(toHash)
-	hBig := new(big.Int).SetBytes(h)
-
-	// If the hash ends up larger than the order of the curve, abort.
-	// Same thing for hash == 0 (as unlikely as that is...).
-	if hBig.Cmp(curve.N) >= 0 {
-		str := fmt.Sprintf("hash of (R || m) too big")
-		return false, schnorrError(ErrSchnorrHashValue, str)
-	}
-	if hBig.Cmp(bigZero) == 0 {
-		str := fmt.Sprintf("hash of (R || m) is zero value")
-		return false, schnorrError(ErrSchnorrHashValue, str)
-	}
-
-	// Convert s to big int.
-	sBig := encodedBytesToBigInt(copyBytes(sigS))
-
-	// We also can't have s greater than the order of the curve.
-	if sBig.Cmp(curve.N) >= 0 {
-		str := fmt.Sprintf("s value is too big")
-		return false, schnorrError(ErrInputValue, str)
-	}
-
-	// r can't be larger than the curve prime.
-	rBig := encodedBytesToBigInt(copyBytes(sigR))
-	if rBig.Cmp(curve.P) == 1 {
-		str := fmt.Sprintf("given R was greater than curve prime")
-		return false, schnorrError(ErrBadSigRNotOnCurve, str)
-	}
-
-	// r' = hQ + sG
-	lx, ly := curve.ScalarMult(pubkey.X(), pubkey.Y(), h)
-	rx, ry := curve.ScalarBaseMult(sigS)
-	rlx, rly := curve.Add(lx, ly, rx, ry)
-
-	if rly.Bit(0) == 1 {
-		str := fmt.Sprintf("calculated R y-value was odd")
-		return false, schnorrError(ErrBadSigRYValue, str)
-	}
-	if !curve.IsOnCurve(rlx, rly) {
-		str := fmt.Sprintf("calculated R point was not on curve")
-		return false, schnorrError(ErrBadSigRNotOnCurve, str)
-	}
-	rlxB := bigIntToEncodedBytes(rlx)
-
-	// r == r' --> valid signature
-	if !bytes.Equal(sigR, rlxB[:]) {
-		str := fmt.Sprintf("calculated R point was not given R")
-		return false, schnorrError(ErrUnequalRValues, str)
-	}
-
-	return true, nil
-}
